Add tests and relocate KeyPaddingUtil
